Guard on delete to only delete if file exists :)

+++ src/index-md/directory-processor/directory-cleaner.ts
@@ -21,6 +21,7 @@ export class DirectoryCleaner extends BaseDirectoryProcessor {
   }
 
   protected deleteFileSync(filePath: string) {
+    if (!this.hasFileAt(filePath)) return;
     fs.removeSync(filePath);
   }
